fix: system tool config
