+++ packages/global/core/app/plugin/utils.ts
@@ -1,6 +1,7 @@
 import { type StoreNodeItemType } from '../../workflow/type/node';
 import { type FlowNodeInputItemType } from '../../workflow/type/io';
 import { FlowNodeTypeEnum } from '../../workflow/node/constant';
+import { PluginSourceEnum } from './constants';
 
 export const getPluginInputsFromStoreNodes = (nodes: StoreNodeItemType[]) => {
   return nodes.find((node) => node.flowNodeType === FlowNodeTypeEnum.pluginInput)?.inputs || [];
@@ -22,3 +23,34 @@ export const getPluginRunContent = ({
   });
   return JSON.stringify(pluginInputsWithValue);
 };
+
+/**
+  plugin id rule:
+  - personal: ObjectId
+  - commercial: commercial-ObjectId
+  - systemtool: systemTool-id
+  (deprecated) community: community-id
+*/
+export function splitCombinePluginId(id: string) {
+  const splitRes = id.split('-');
+  if (splitRes.length === 1) {
+    // app id
+    return {
+      source: PluginSourceEnum.personal,
+      pluginId: id
+    };
+  }
+
+  const [source, pluginId] = id.split('-') as [PluginSourceEnum, string | undefined];
+  if (!source || !pluginId) throw new Error('pluginId not found');
+
+  // 兼容4.10.0 之前的插件
+  if (source === 'community' || id === 'commercial-dalle3') {
+    return {
+      source: PluginSourceEnum.systemTool,
+      pluginId: `${PluginSourceEnum.systemTool}-${pluginId}`
+    };
+  }
+
+  return { source, pluginId: id };
+}
