+++ projects/app/src/pages/api/core/app/plugin/getPreviewNode.ts
@@ -3,16 +3,14 @@
  */
 import { PluginSourceEnum } from '@fastgpt/global/core/app/plugin/constants';
 import { ReadPermissionVal } from '@fastgpt/global/support/permission/constant';
-import {
-  getChildAppPreviewNode,
-  splitCombinePluginId
-} from '@fastgpt/service/core/app/plugin/controller';
+import { getChildAppPreviewNode } from '@fastgpt/service/core/app/plugin/controller';
 import { type FlowNodeTemplateType } from '@fastgpt/global/core/workflow/type/node.d';
 import { NextAPI } from '@/service/middleware/entry';
 import { type ApiRequestProps } from '@fastgpt/service/type/next';
 import { authApp } from '@fastgpt/service/support/permission/app/auth';
 import type { NextApiResponse } from 'next';
 import { getLocale } from '@fastgpt/service/common/middle/i18n';
+import { splitCombinePluginId } from '@fastgpt/global/core/app/plugin/utils';
 
 export type GetPreviewNodeQuery = { appId: string; versionId?: string };
